+++ CHANGELOG.md
@@ -11,8 +11,65 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixes
 
+### Breaking
+
+## [2.47.0] - 2023-11-26
+
+### Features
+
+- program: accelerated idle update for users with <$1000 equity
+
+### Fixes
+
+### Breaking
+
+## [2.46.0] - 2023-11-24
+
+### Features
+
+- program: fill asset weight between init and maintenance ([#713](https://github.com/drift-labs/protocol-v2/pull/713))
+- program: if order reduces maker position, check maintenance margin requirement ([#714](https://github.com/drift-labs/protocol-v2/pull/714))
+
+### Fixes
+
+### Breaking
+
+## [2.45.0] - 2023-11-22
+
+### Features
+
+- program: trigger limits cant make if limit crosses trigger ([#707](https://github.com/drift-labs/protocol-v2/pull/702))
+
+### Fixes
+
+- sdk: fix vamm L2asks by using askAmm ([#708](https://github.com/drift-labs/protocol-v2/pull/708))
+- program: add max_number_of_sub_accounts onto state account ([#710](https://github.com/drift-labs/protocol-v2/pull/710))
+
+### Breaking
+
+## [2.44.0] - 2023-11-17
+
+### Features
+
+### Fixes
+
+- program: exhaustively search for referrer account on fill ([#702](https://github.com/drift-labs/protocol-v2/pull/702))
+
+## [2.43.0] - 2023-11-16
+
+### Features
+
+- program: accelerate liquidations for tiny accounts ([#698](https://github.com/drift-labs/protocol-v2/pull/698))
+- program: boost max sub accounts to 20k
+
+### Fixes
+
+- program: allow amm to fill step size ([#672](https://github.com/drift-labs/protocol-v2/pull/672))
+- program: add add update_liquidation_margin_buffer_ratio ([#695](https://github.com/drift-labs/protocol-v2/pull/695))
 - program: account for fee pool when settling positive pnl ([#687](https://github.com/drift-labs/protocol-v2/pull/687))
 - sdk: fix bug which incorrectly calculated leverage after trade for a market with no position but short orders open
+- sdk: fix bug in modifying an order that previously had auction params to a non-auction order
+- sdk: add delta to calculateDepositRate function
 
 ### Breaking